conditionally rendered the data for the three tabs on the settings page on click.
