+++ client/src/containers/Private/Settings/Components/GasComponentData.js
@@ -129,39 +129,7 @@ const GasComponentData = () => {
         >
           <EditIcon />
         </IconButton>
-      </ListItem>
-      <Divider />
-
-      {/* Username  */}
-
-      <ListItem style={{ marginTop: "10px", marginBottom: "10px" }}>
-        <ListItemText primary="Username:" />
-        <ListItemText primary={`${globalState.user.username}`} />
-        <IconButton
-          className="editButton"
-          onClick={() => handleOpen("username")}
-          edge="end"
-          aria-label="edit"
-        >
-          <EditIcon />
-        </IconButton>
-      </ListItem>
-      <Divider />
-
-      {/* Password  */}
-
-      <ListItem style={{ marginTop: "10px", marginBottom: "10px" }}>
-        <ListItemText primary="Password:" />
-        <ListItemText primary="**********" />
-        <IconButton
-          className="editButton"
-          onClick={() => handleOpen("password")}
-          edge="end"
-          aria-label="edit"
-        >
-          <EditIcon />
-        </IconButton>
-      </ListItem>
+      </ListItem>    
       <Divider />
     </List>
   );
